Add Scatter3D Visualization Operator (#2346)
This pull request introduces the Satter3D Plot viz operator.

A Scatter3D Plot, often referred to as a Satter3D Chart, is a type of
data visualization primarily used to show a 3D Scatter Chart based on a
set of 3D datasets. This operator allows the user to select 3 groups of
data as x, y, and z, then the operator will render the visualization.

For implementation details, this operator is a python-native operator.
It accepts the user-specified data and generate corresponding python
script. The figure is created with plotly.graph_objects.Figure. More
details about the plotly Scatter3D chart drawing is available
[here](https://plotly.com/python/reference/scatter3d/).

Here is the demo:

![Record](https://github.com/Texera/texera/assets/62498853/754bfbaa-0985-4e6b-8133-5c43ae3f313e)

---------

Co-authored-by: Xiaozhen Liu <xiaozl3@uci.edu>