Add materializations without LogicalPlan (#2377)
Previously, when adding a materialization, we rely on the
LogicalOperators and the change will need information from LogicalPlan
(a.k.a, the schema). Now since schemas are moved onto ports, we will
gradually push down the adding materializations as a PhysicalPlan change
only. This PR removes the LogicalPlan access during adding
materializations. In the future we can also remove dependency to Logical
Operators during this process (after we push done storage).