Apply schema check on operator output port (#2394)
This PR introduces the `TupleLike` concept to Scala Amber, replacing the
traditional `Tuple` output of operators. The Tuple implementation
requires a schema for construction. By adopting `TupleLike`, this PR
pushes the schema requirement to the output port.

Now developers can write the following code in the operator logic to
construct a tuple:
```
TupleLike(1,"a", "b")
```
or
```
TupleLike("col1" -> 1, "col2" -> "a", "col3" -> "b")
```