Fix operators showing "Running" state after workflow completion (#2411)
This PR addresses an issue where operators are not marked as finished
upon the completion of a workflow. The underlying problem stems from the
premature sending of PortComplete messages before WorkerComplete
messages. Specifically, within the PortCompleteHandler, the current
region is marked as completed. Consequently, when WorkerCompleteHandler
executes, it attempts to query stats for the currently running region,
which has already been marked as completed due to the prior PortComplete
call. To resolve this issue, this PR introduces an additional call to
querystats before a region is marked as completed, ensuring that the
operators are correctly finished after workflow completion.