Remove Schema usage from OperatorExecutor (#2419)
This PR removes the schema usage from all operator executors except for
source operators. This decouples the life cycle of an executor from the
life cycle of schemas on ports.
- For non-source operators, an executor should not depend on the input
schema or output schema as they are dynamic to the data stream.
- For source operators, currently, we rely on schema to do type casting
so that source operators generate data with more user-friendly types. A
temporary workaround is applied to let source operators fetch output
schema during their open() stage.