Refactor Aggregate operator (#2438)
This PR refactors Aggregate operator:
1. Removed `PartialAggregateOpExec` and `FinalAggregtateOpExec` by
combining them into the same executor: `AggregateOpExec`, unified their
input and output apis.
2. Make internal ports have the correct aggregate schema instead of
key-value schema which omits the actual data information, also this
unifies the output schema of partial and final aggregates.
3. Cleaned the codebase by merging SpecializedAggregateOpDesc into
`AggregateOpDesc`, as it is the only implementation now.
`AggregateOpDesc` is not an abstract class anymore. I have also moved it
to the package that aligned with other operators.