Introducing Java UDF (#2448)
The PR introduces the Java User Defined function to enable custom Java code compilation in Texera. 
Java UDF ensures runtime compilation of Java code and also compilation without saving source code in memory.
The source code is sent from the web server to the controller and the controller sends the source code to the workers. Then inside the workers, the java code is compiled and the compiled class is instantiated and sent to the Executor. The following figure shows the phenomenon:
 
<img src="https://github.com/Texera/texera/assets/43316759/1a03a190-8ab7-4719-8cef-4e1ca470b436" width="550" />


Each worker gets the source code, passes the code to a JavaFileObject, and creates an in-memory file which is sent to the compiler to compile. Next, to get the compiled bytes, another JavaFileObject is passed to the file manager of the compiler. Finally, the compiled bytes are sent to the custom classloader to load the class.

<img src="https://github.com/Texera/texera/assets/43316759/9d3f0a22-6f87-4351-9e0f-8b515271de2d" width="500" />


If there are multiple workers, they follow the same procedure concurrently.

<img src="https://github.com/Texera/texera/assets/43316759/9c710a68-28a0-4ad0-8b69-ed5e5ab64dbb" width="500" />


We can see the demonstration in the following:
![ezgif-7-d7ebd5a108](https://github.com/Texera/texera/assets/43316759/d3b828be-9536-4d49-9ebc-2a1a7bf51288)

---------

Co-authored-by: Yicong Huang <17627829+Yicong-Huang@users.noreply.github.com>