Make Property Editor and Minimap Panel Resizable, Draggable, and Collapsible (#2435)
This PR makes the property panel draggable, resizable, and also
colllapsible. The property efitor panel will not be shown when users
deselect operators, but it will be displayed when users select any
operator.

The Minimap panel is fixed to the bottom right of the page and can be
collapsed by clicking the minimap icon.

The location and status of each panel will be memorized after refreshing
the page.

This PR also replaces ng-dynamic with Angular native functions.


https://github.com/Texera/texera/assets/86388854/6b73a017-4949-4220-a7c1-6b4017fbdbe3

---------

Co-authored-by: MiuMiuMiue <yuanzhe2001@gmail.com>
Co-authored-by: mengw15 <simonmengwang123@gmail.com>
Co-authored-by: Xinyuan Lin <xinyual3@uci.edu>