Fix incorrect execution state aggregation for multi-region workflows (#2585)
If the workflow has 2 regions and operator A is shared between regions,
then the aggregated workflow execution state also considers the
execution state of A in the previous region. This behavior prevents the
workflow from being paused on the frontend for multi-region workflows.

This PR removes the operators in a completed region so that the same
operator will not be considered twice.